requirements.txt
fastapi
uvicorn[standard]
websocket-client
scikit-learn
numpy
pandas
loguru
